Showing insufficient privileges if not admin for dashboard

+++ src/components/AdminDashboard/AdminDashboard.js
@@ -6,42 +6,53 @@ import PropTypes from 'prop-types';
 
 /* Admin dashboard component for deleting and creating forums */
 const AdminDashboard = (props) => {
-  const { forums, deleteForum, deletingForumId } = props;
-  return (
-    <div className='dashboard'>
-      <div className='dashboard__forums'>
-        <div className='dashboard__title'>
-          <h2>Forums</h2>
-        </div>
-        {forums &&
-          forums.map((forum) => {
-            let deleteForumContent = null;
-            if (deletingForumId === forum._id) {
-              deleteForumContent = (
-                <span className='dashboard__delete'>Deleting...</span>
-              );
-            } else {
-              deleteForumContent = (
-                <div
-                  className='dashboard__delete'
-                  onClick={() => deleteForum(forum._id)}
-                >
-                  <FaTrashAlt className='dashboard__delete-icon' />
-                  <span>Delete</span>
+  const { user, forums, deleteForum, deletingForumId } = props;
+
+  let dashboardContent = null;
+  if (user && user.role === 'admin') {
+    dashboardContent = (
+      <React.Fragment>
+        <div className='dashboard__forums'>
+          <div className='dashboard__title'>
+            <h2>Forums</h2>
+          </div>
+          {forums &&
+            forums.map((forum) => {
+              let deleteForumContent = null;
+              if (deletingForumId === forum._id) {
+                deleteForumContent = (
+                  <span className='dashboard__delete'>Deleting...</span>
+                );
+              } else {
+                deleteForumContent = (
+                  <div
+                    className='dashboard__delete'
+                    onClick={() => deleteForum(forum._id)}
+                  >
+                    <FaTrashAlt className='dashboard__delete-icon' />
+                    <span>Delete</span>
+                  </div>
+                );
+              }
+              return (
+                <div key={forum.name} className='dashboard__forum'>
+                  <div className='dashboard__forum-name'>{forum.name}</div>
+                  {deleteForumContent}
                 </div>
               );
-            }
-            return (
-              <div key={forum.name} className='dashboard__forum'>
-                <div className='dashboard__forum-name'>{forum.name}</div>
-                {deleteForumContent}
-              </div>
-            );
-          })}
+            })}
+        </div>
+        <CreateForumContainer />
+      </React.Fragment>
+    );
+  } else {
+    dashboardContent = (
+      <div className='dashboard__privileges'>
+        You do not have sufficient privileges to view this page.
       </div>
-      <CreateForumContainer />
-    </div>
-  );
+    );
+  }
+  return <div className='dashboard'>{dashboardContent}</div>;
 };
 
 AdminDashboard.propTypes = {
